Refactor and enhance codebase for improved performance and readability
- Updated integer declarations to use `integer(int32)` for consistency and to ensure compatibility across different platforms.
- Introduced new module `core_memory` to encapsulate memory allocation and deallocation functionalities.
- Added `core_validation` module for enhanced error checking and validation functionalities.
- Refactored various subroutines in `core_types_variable` to improve clarity and maintainability, including the addition of new methods for managing previous variable states.
- Improved documentation throughout the codebase, translating comments from Japanese to English for better accessibility.
- Changed several subroutines from pure elemental to regular subroutines to enhance flexibility in usage.
- Updated pointer declarations to use `pointer` attributes for dynamic memory management.
- Enhanced the handling of history data in `thermal` and `hydraulic` modules for better performance in simulations.

+++ src/core/interop/interop_utils.F90
@@ -1,4 +1,5 @@
 module core_c_utils
+    use, intrinsic :: iso_fortran_env
     use, intrinsic :: iso_c_binding, only: c_ptr, c_f_pointer, c_char, c_null_char
 
     implicit none
@@ -16,9 +17,9 @@ contains
         character(:), allocatable :: f_string
 
         character(kind=c_char), pointer :: c_char_array_ptr(:)
-        integer :: i, length
+        integer(int32) :: i, length
         ! 非常に大きなサイズ（ここでは2^30）を指定
-        integer, parameter :: large_size = 2**30
+        integer(int32), parameter :: large_size = 2**30
 
         ! --- 1. 長さを決定する ---
         if (.not. c_associated(c_ptr_in)) then
